pyproject.toml
[project]
name = "referalbot"
version = "0.1.0"
description = ""
authors = [
    {name = "reddmonchick",email = "reddmonchick375@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "referalbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
